ProductApi/Backend/pages/requirements.txt
asgiref==3.2.10
dataclasses==0.7
Django==3.0.5
django-extra-fields==3.0.2
djangorestframework==3.12.1
djongo==1.3.3
Pillow==8.0.1
pkg-resources==0.0.0
psycopg2==2.8.6
pymongo==3.11.0
pytz==2020.1
sqlparse==0.2.4
